% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{auto_anno}
\alias{auto_anno}
\title{Generate annotations from data}
\usage{
auto_anno(data, mode, limit_universe = TRUE, save = NULL)
}
\arguments{
\item{data}{\code{\link[E.PATH:database]{E.PATH::database}} or output of
\code{\link{import_database}}}

\item{mode}{"GO", "KEGG" or "MeSH"}

\item{limit_universe}{limit_universe limit universe to genes contained in \code{gs_genes}}

\item{save}{optional: path to save annotations as \code{.csv}}
}
\value{
tibble: raw annotations
}
\description{
Generate annotations from data
}
\examples{
data_path <- system.file("extdata/ex_data.csv", package="E.PAGE")
data <- import_database(data_path, ",", FALSE, c(2, 4), 0)

auto_anno(data, "GO", limit_universe = FALSE, save = "anno.csv")
}
