% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{compute_sets}
\alias{compute_sets}
\title{Compute overlap enrichment on E.PATH gene set categories}
\usage{
compute_sets(
  input,
  universe = NULL,
  info_from = "database",
  anno_opts = "file",
  categories = FALSE,
  organisms = FALSE,
  org_db = NULL
)
}
\arguments{
\item{input}{output of \code{\link{process_input_text}} or
\code{\link{process_input_seurat}}}

\item{universe}{number of genes in universe}

\item{info_from}{optional: \code{"annotation"} or (default) \code{"database"}}

\item{anno_opts}{optional: \code{"name"} for gene set names, \code{"syms"}
for gene set symbols, \code{"info"} for gene set descriptions,
\code{"auto"} for automatically generated annotations and/or (default)
\code{"file"} for manual annotations}

\item{categories}{optional: categories to include; default all}

\item{organisms}{optional: organisms to include; default all}

\item{org_db}{optional: GO organism database e.g.
\code{\link[org.Hs.eg.db:org.Hs.eg.db]{org.Hs.eg.db::org.Hs.eg.db}}}
}
\value{
list: categories and output of \code{\link{compute}}
}
\description{
Compute overlap enrichment on E.PATH gene set categories
}
\examples{
input <- process_input_text("FCN1 0.1 FTL 0.8 CLU 0.05")
results <- compute_sets(input)
}
