% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_annotations}
\alias{import_annotations}
\title{Import delimited annotation file into E.PAGE-specific format}
\usage{
import_annotations(path, delim, header, content = NULL, info = NULL)
}
\arguments{
\item{path}{path to file}

\item{delim}{file delimiter, usually \code{","} or \code{"\\t"}}

\item{header}{whether file contains header}

\item{content}{vector: first and last column of content; default 2 to end}

\item{info}{column containing information; default none}
}
\value{
E.PAGE-specific imported annotations
}
\description{
Import delimited annotation file into E.PAGE-specific format
}
\examples{
anno_path <- system.file("extdata", "ex_anno.csv", package = "E.PAGE")
anno <- import_annotations(anno_path, ",", TRUE, c(2, 4), 5)
}
