% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{synthesize_go_anno}
\alias{synthesize_go_anno}
\title{Synthesize annotation file from processed data}
\usage{
synthesize_go_anno(
  data,
  org_db = org.Hs.eg.db,
  extra_args = list(keyType = "SYMBOL", ont = "ALL"),
  limit_universe = TRUE,
  save = NULL
)
}
\arguments{
\item{data}{\code{\link[E.PATH:database]{E.PATH::database}} or output of
\code{\link{import_database}}}

\item{org_db}{GO organism database e.g.
\code{\link[org.Hs.eg.db:org.Hs.eg.db]{org.Hs.eg.db::org.Hs.eg.db}}}

\item{extra_args}{extra arguments for
\code{\link[clusterProfiler:enrichGO]{clusterProfiler::enrichGO}}}

\item{limit_universe}{limit universe to genes contained in \code{gs_genes}}

\item{save}{optional: path to save annotations as \code{.csv}}
}
\value{
tibble: raw annotations
}
\description{
Synthesize annotation file from processed data
}
\examples{
data_path <- system.file("extdata/ex_data.csv", package="E.PAGE")
data <- import_database(data_path, ",", FALSE, c(2, 4), 0)

synthesize_go_anno(data, limit_universe = FALSE, save = "anno.csv")
}
