% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{synthesize_go_anno}
\alias{synthesize_go_anno}
\title{Synthesize annotation file from processed data}
\usage{
synthesize_go_anno(
  gs_genes,
  org_db = org.Hs.eg.db,
  extra_args = list(keyType = "SYMBOL", ont = "ALL"),
  limit_universe = TRUE,
  save = NULL
)
}
\arguments{
\item{gs_genes}{value from \code{\link{process_database}}}

\item{org_db}{GO organism database e.g.
\code{\link[org.Hs.eg.db:org.Hs.eg.db]{org.Hs.eg.db::org.Hs.eg.db}}}

\item{extra_args}{extra arguments for
\code{\link[clusterProfiler:enrichGO]{clusterProfiler::enrichGO}}}

\item{limit_universe}{limit universe to genes contained in \code{gs_genes}}

\item{save}{optional: path to save annotations as \code{.csv}}
}
\value{
tibble: raw annotations
}
\description{
Synthesize annotation file from processed data
}
\examples{
gs_genes <- E.PATH::database$gs_genes
synthesize_go_anno(gs_genes[0:2], limit_universe = FALSE)
}
